mp: Refactor code structure
