+++ scripts/mp/_env.js
@@ -1,0 +1,5 @@
+exports.IF = {
+  unit: 'rpx',
+  useRemote: false,
+  ctx: null
+}
