+++ scripts/mp/_genInjectCSS.js
@@ -6,13 +6,14 @@ function genInjectCSS() {
   let { bgc, gft = 'inherit' } = IF.ctx.Config.setting
 
   let IARoad = getPath('style/inject.less')
-  let bgContent = `html,body { background-color: ${bgc};}\n.U-unit { font-family: ${gft};}\n`
+  let bgContent = `.page { background-color: ${bgc}; }\n.U-unit { font-family: ${gft};}\n`
 
+  // If you want to load web fonts dynamically, the file address needs to be the download type.
+  // https://developers.weixin.qq.com/miniprogram/dev/api/ui/font/wx.loadFontFace.html
   let fontContent = Object.keys(FontList).map(name => {
-    let url = IF.useRemote ? `${FontCDN}fonts/${name}.woff2` : `/assets/${name}.woff`
+    let url = `${FontCDN}fonts/${name}.woff`
     return `@font-face {font-family: '${name}';src:url('${url}')};`
   }).join('\n')
-  
 
   writeIn(IARoad, format(bgContent + fontContent))
 }
