+++ temps/mp/package.json
@@ -27,7 +27,7 @@
     "dev:quickapp": "npm run build:quickapp -- --watch"
   },
   "browserslist": [
-    "last 10 versions",
+    "last 3 versions",
     "Android >= 4.1",
     "ios >= 8"
   ],
@@ -35,9 +35,10 @@
   "license": "MIT",
   "dependencies": {
     "@babel/runtime": "^7.7.7",
-    "@tarojs/components": "3.3.14",
-    "@tarojs/runtime": "3.3.14",
-    "@tarojs/taro": "3.3.14",
+    "@tarojs/runtime": "3.4.8",
+    "@tarojs/taro": "3.4.8",
+    "@tarojs/components": "3.4.8",
+    "@tarojs/plugin-framework-vue2": "3.4.8",
     "blueimp-md5": "^2.18.0",
     "lodash.clonedeep": "^4.5.0",
     "pubsub-js": "^1.9.3",
@@ -48,14 +49,14 @@
   },
   "devDependencies": {
     "@types/webpack-env": "^1.13.6",
-    "@tarojs/mini-runner": "3.3.14",
-    "@tarojs/webpack-runner": "3.3.14",
+    "@tarojs/mini-runner": "3.4.8",
+    "@tarojs/webpack-runner": "3.4.8",
     "@babel/core": "^7.8.0",
-    "babel-preset-taro": "3.3.14",
+    "babel-preset-taro": "3.4.8",
     "vue-loader": "^15.9.2",
     "eslint-plugin-vue": "^6.x",
-    "eslint-config-taro": "3.3.14",
-    "eslint": "^6.8.0",
+    "eslint-config-taro": "3.4.8",
+    "eslint": "^8.12.0",
     "stylelint": "9.3.0"
   }
 }
